requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==37.0.3
daphne==3.0.2
Django==3.2.12
django-ckeditor==6.2.0
django-environ==0.8.1
django-js-asset==2.0.0
djangorestframework==3.13.1
geographiclib==1.52
geopy==2.2.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
Pillow==8.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pytz==2021.3
redis==3.5.3
requests==2.28.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
twilio==7.9.3
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.9
zope.interface==5.4.0
